Moving IrLazyImMap utility class into org.enso.scala.wrapper.ScalaConversions
